jericho/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Arden Chew <achew@grailbio.com>"]
packages = [{ include = "**/*" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
sqlmodel = "^0.0.8"
uvicorn = "^0.23.2"
psycopg2 = "^2.9.7"
requests = "^2.31.0"
pytest = "^7.4.0"
pytest-postgresql = "^5.0.0"
httpx = "^0.24.1"
pyyaml = "^6.0.1"
pylint = "^2.17.5"
isort = "^5.12.0"
black = "^23.9.0"
flake8 = "^6.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^41.0.4"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
